refactor: creating the save method in the ComputerRepository class
